apps/api/requirements.txt
# LiDAR Processing Libraries
laspy>=2.4.0
laszip>=0.2.0
numpy>=1.24.0
scipy>=1.10.0
ezdxf>=1.1.0
pyproj>=3.6.0

# Open3D (for point cloud processing)
open3d>=0.17.0

# FastAPI and server
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1

# Azure services
azure-storage-blob==12.19.0

# MongoDB (Cosmos DB)
motor==3.3.2
pymongo==4.6.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Development tools
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2